Pickup, Item, ItemSlot Update
Pickup now checks to see if there is any instance of the proposed,
stackable object. If there is, it increments that stack. Otherwise, it
keeps looking and if it doesn't find an unfilled stack, it creates a new
one. Item now has a "stackable" value, something that should be inherent
in Items. Item slot now has a getter for its maxStackSize.